Implementado tratador de erro, Exception de validação

+++ src/main/java/br/com/alura/orcamento_domestico/service/MetaOrcamentoService.java
@@ -1,8 +1,8 @@
 package br.com.alura.orcamento_domestico.service;
 
 import br.com.alura.orcamento_domestico.dto.metasOrcamentoDTO.CadastroMetasOrcamentoDTO;
-import br.com.alura.orcamento_domestico.dto.metasOrcamentoDTO.ListaCadastroMetasOrcamentoDTO;
 import br.com.alura.orcamento_domestico.dto.metasOrcamentoDTO.DetalheMetaOrcamentoDTO;
+import br.com.alura.orcamento_domestico.exception.ValidacaoException;
 import br.com.alura.orcamento_domestico.model.MetaOrcamento;
 import br.com.alura.orcamento_domestico.repository.MetaOrcamentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,8 +16,13 @@ public class MetaOrcamentoService {
     @Autowired
     private MetaOrcamentoRepository metaOrcamentoRepository;
 
-    public List<DetalheMetaOrcamentoDTO> cadastrar(ListaCadastroMetasOrcamentoDTO dados) {
-        List<MetaOrcamento> metasOrcamento = converterListaDtoParaMetas(dados.metas());
+    public List<DetalheMetaOrcamentoDTO> cadastrar(List<CadastroMetasOrcamentoDTO> dados) throws ValidacaoException{
+        double somaPorcentagem = dados.stream().map(CadastroMetasOrcamentoDTO::porcentagem).reduce( 0.0, Double::sum);
+        if (somaPorcentagem != 100.0){
+            throw new ValidacaoException("Soma das metas deve ser igual a 100%");
+        }
+
+        List<MetaOrcamento> metasOrcamento = converterListaDtoParaMetas(dados);
         metaOrcamentoRepository.deleteAll();
         metasOrcamento.forEach(metaOrcamento -> metaOrcamentoRepository.save(metaOrcamento));
         return converterListaMetasParaDto(metaOrcamentoRepository.findAll());
